chore: add comment on module factory
